Move config package
