+++ internal/docker/legacydocker/legacydocker_test.go
@@ -14,8 +14,8 @@ import (
 
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
-	"github.com/saucelabs/saucectl/cli/config"
 	"github.com/saucelabs/saucectl/cli/mocks"
+	"github.com/saucelabs/saucectl/internal/config"
 	"github.com/stretchr/testify/assert"
 	"gotest.tools/v3/fs"
 )
