+++ internal/fleet/fleet_test.go
@@ -2,7 +2,7 @@ package fleet
 
 import (
 	"context"
-	"github.com/saucelabs/saucectl/cli/config"
+	"github.com/saucelabs/saucectl/internal/config"
 	"gotest.tools/v3/fs"
 	"strings"
 	"testing"
@@ -70,7 +70,7 @@ func TestRegister(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := Register(tt.args.ctx, tt.args.seq, "",tt.args.testFiles, tt.args.suites)
+			got, err := Register(tt.args.ctx, tt.args.seq, "", tt.args.testFiles, tt.args.suites)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("Register() error = %v, wantErr %v", err, tt.wantErr)
 				return
